src/pumps-agent/ai-foundry/requirements.txt
azure-ai-projects --pre
azure-identity
azure-mgmt-cognitiveservices
azure-search-documents
openai
python-dotenv